exercise

+++ src/hust/soict/dsai/aims/media/Disc.java
@@ -1,7 +1,7 @@
 package hust.soict.dsai.aims.media;
 
 public class Disc extends Media {
-    protected float length;
+    protected int length;
     protected String director;
     public Disc(){};
     public Disc(float length, String director) {
